fix duel::write_buffer should use size_t

+++ duel.cpp
@@ -121,7 +121,7 @@ void duel::restore_assumes() {
 		pcard->assume_type = 0;
 	assumes.clear();
 }
-void duel::write_buffer(const void* data, int size) {
+void duel::write_buffer(const void* data, size_t size) {
 	vector_write_block(message_buffer, data, size);
 }
 void duel::write_buffer32(uint32_t value) {
